Pulling out minibatch functionality into new python file.
